feat: async request in onLoad when immediate is true, you can access the options
