+++ .editorconfig
@@ -1,12 +1,8 @@
 root = true
 
-[{*.js,*.css,*.html}]
+[*]
 indent_style = space
-indent_size = 4
+indent_size = 2
 end_of_line = lf
 charset = utf-8
 insert_final_newline = true
-
-[{package.json,.*rc,*.yml}]
-indent_style = space
-indent_size = 2
