+++ .vscode/settings.json
@@ -7,7 +7,7 @@
 	},
 	"editor.suggestSelection": "recentlyUsedByPrefix", // 自动补全
 	"editor.codeActionsOnSave": {
-		"source.organizeImports": true,
-		"source.fixAll.eslint": true // 修复所有代码
+		"source.organizeImports": "explicit",
+		"source.fixAll.eslint": "explicit"
 	}
 }
