--- commitlint.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['@commitlint/config-conventional']
-};
