+++ src/index.ts
@@ -1,4 +1,7 @@
+import { onLoad } from '@dcloudio/uni-app';
+import { EffectRequestParams } from 'alova';
 import VueHook from 'alova/vue';
+import { WatchSource } from 'vue';
 import { AdapterUniappOptions } from '../typings';
 import requestAdapter from './requestAdapter';
 import storageAdapter from './storageAdapter';
@@ -8,7 +11,19 @@ export { default as uniappStorageAdapter } from './storageAdapter';
 
 export default function AdapterUniapp({ mockRequest }: AdapterUniappOptions = {}) {
 	return {
-		statesHook: VueHook,
+		statesHook: {
+			...VueHook,
+			effectRequest(effectRequestParams: EffectRequestParams<WatchSource>) {
+				const handler = effectRequestParams.handler;
+				effectRequestParams.handler = () => {
+					// 延迟10ms让页面中的onLoad先执行
+					onLoad(() => {
+						setTimeout(handler, 10);
+					});
+				};
+				return VueHook.effectRequest(effectRequestParams);
+			}
+		},
 		requestAdapter: mockRequest || requestAdapter,
 		storageAdapter
 	};
